requirements.txt
cupy-cuda102==10.2.0
easydict==1.9
jupyter==1.0.0
kornia==0.6.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
numba==0.56.2
numpy==1.21.5
opencv-python==4.5.5.62
Pillow==9.2.0
protobuf==3.20.1
PyYAML==6.0
scipy==1.7.3
tensorboardX==2.5.1
termcolor==1.1.0
termcolor-whl==1.1.2
timm==0.6.7
torch==1.12.1
torchvision==0.13.1
tqdm==4.62.3
wandb==0.12.10
pandas==1.3.5
streamlit-option-menu
streamlit